Change `Result<()>` to `ProgramResult`
An alias of `Result<T>` is generated from the `errors` macro. This means
that any examples that don't use the `errors` macro will fail at compile
time.

For a beginner to Rust, this can lead to a confusing error that's hard to
debug.

My suggestion is to standardize on `ProgramResult` which is also used in
the examples section of the Anchor repo.